Add new composable useSearch
